+++ com.unity.visualeffectgraph/Editor/Models/VFXUI.cs
@@ -93,7 +93,6 @@ namespace UnityEditor.VFX
 
         public GroupInfo[] groupInfos;
         public StickyNoteInfo[] stickyNoteInfos;
-        public List<SystemInfo> systemInfos;
 
         [Serializable]
         public struct CategoryInfo
@@ -105,60 +104,6 @@ namespace UnityEditor.VFX
         public List<CategoryInfo> categories;
 
         public Rect uiBounds;
-
-        public string GetNameOfSystem(IEnumerable<VFXContext> contexts)
-        {
-            if(systemInfos != null)
-            {
-                foreach(var context in contexts)
-                {
-                    var system = systemInfos.Find(t => t.contexts.Contains(context));
-                    if (system != null)
-                        return system.title;
-                }
-            }
-            return string.Empty;
-        }
-
-        public void SetNameOfSystem(IEnumerable<VFXContext> contexts, string name)
-        {
-            if( systemInfos == null)
-            {
-                systemInfos = new List<SystemInfo>();
-            }
-            foreach (var context in contexts)
-            {
-                var system = systemInfos.Find(t => t.contexts.Contains(context));
-                if (system != null)
-                {
-                    system.contexts = contexts.ToArray();
-                    system.title = name;
-
-                    // we found a matching system, clean all other of these contexts
-                    foreach( var s in systemInfos)
-                    {
-                        if( s != system)
-                        {
-                            if( s.contexts.Intersect(contexts) != null)
-                            {
-                                s.contexts = s.contexts.Except(contexts).ToArray();
-                            }
-                        }
-                    }
-
-                    if( string.IsNullOrEmpty(name))
-                    {
-                        systemInfos.Remove(system);
-                    }
-                    return;
-                }
-            }
-            if( ! string.IsNullOrEmpty(name) )
-            {
-                // no system contains any of the contexts. Add a new one.
-                systemInfos.Add(new SystemInfo() { contexts = contexts.ToArray(), title = name });
-            }
-        }
 
         public void Sanitize(VFXGraph graph)
         {